initial commit : custom and global exceptional Handling

+++ src/main/java/com/stackroute/muzixapp/service/MuzixServiceImpl.java
@@ -1,11 +1,16 @@
 package com.stackroute.muzixapp.service;
 
 import com.stackroute.muzixapp.domain.Muzix;
+import com.stackroute.muzixapp.exceptions.TrackAlreadyExistsException;
+import com.stackroute.muzixapp.exceptions.TrackNotFoundException;
 import com.stackroute.muzixapp.repository.MuzixRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import javax.sound.midi.Track;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class MuzixServiceImpl implements MuzixService {
@@ -16,38 +21,58 @@ public class MuzixServiceImpl implements MuzixService {
         this.muzixRepository=userRepository;
     }
     @Override
-    public Muzix saveUser(Muzix muzix) {
+    public Muzix saveUser(Muzix muzix) throws TrackAlreadyExistsException {
+        if(muzixRepository.existsById(muzix.getTrackId())){
+            throw new TrackAlreadyExistsException("Track is already present");
+        }
         Muzix savedUser=muzixRepository.save(muzix);
+        if(savedUser==null){
+            throw new TrackAlreadyExistsException("Track was already pressent");
+        }
         return savedUser;
     }
 
     @Override
-    public List<Muzix> getAllUsers() {
+    public List<Muzix> getAllUsers()  {
         return muzixRepository.findAll();
     }
 
     @Override
-    public List<Muzix> deleteById(int trackId)
+    public List<Muzix> deleteById(int trackId) throws TrackNotFoundException
     {
+        if(!muzixRepository.existsById(trackId)){
+            throw new TrackNotFoundException("track is not present");
+        }
         muzixRepository.deleteById(trackId);
         return muzixRepository.findAll();
 
     }
     @Override
-    public void findById(int trackId){
-        muzixRepository.findById(trackId);
-        return;
+    public boolean findById(int  trackId) throws TrackNotFoundException{
+      if( !muzixRepository.existsById(trackId)){
+          throw new TrackNotFoundException("track not found");
+        }
+         muzixRepository.findById(trackId);
+        return true;
 
     }
     @Override
-    public List<Muzix> updateUsers(Muzix muzix,int trackId){
-        muzixRepository.findById(trackId);
-        Muzix updatedUser=muzixRepository.save(muzix);
-        return muzixRepository.findAll();
+    public ResponseEntity<Object> updateMuzix(Muzix muzix, int id)throws TrackNotFoundException {
+        Optional<Muzix> trackOptional = muzixRepository.findById(id);
+        if (!trackOptional.isPresent()) {
+            throw new TrackNotFoundException("track is not present");
 
+        }
+
+        muzix.setTrackId(id);
+        muzixRepository.save(muzix);
+        return ResponseEntity.noContent().build();
     }
     @Override
-    public List<Muzix> getMuzixByName(String name){
+    public List<Muzix> getMuzixByName1(String name)throws TrackNotFoundException{
+        if(muzixRepository.getMuzixByName(name).isEmpty()){
+            throw new TrackNotFoundException((" track not found"));
+        }
          return muzixRepository.getMuzixByName(name);
 
     }
